+++ GLImp/Textures/Subimage.cs
@@ -21,39 +21,39 @@ namespace GLImp {
 			this.Yoffset = y;
 		}
 
-		private float Left {
+		private double Left {
 			get {
 				return (Xoffset * 1.0f) / texture.Width;
 			}
 		}
 
-		private float Right {
+		private double Right {
 			get {
 				return ((Xoffset + 1) * 1.0f) / texture.Width;
 			}
 		}
 
-		private float Top {
+		private double Top {
 			get {
 				return (Yoffset * 1.0f) / texture.Height;
 			}
 		}
 
-		private float Bottom {
+		private double Bottom {
 			get {
 				return ((Yoffset + 1) * 1.0f) / texture.Height;
 			}
 		}
 
-		public override void Draw(float x, float y) {
+		public override void Draw(double x, double y) {
 			this.Draw(x, y, this.Width, this.Height);
 		}
 
-		public override void Draw(float X, float Y, float Width, float Height) {
-			float X1 = X;
-			float Y1 = Y;
-			float X2 = X + Width;
-			float Y2 = Y + Height;
+		public override void Draw(double X, double Y, double Width, double Height) {
+			double X1 = X;
+			double Y1 = Y;
+			double X2 = X + Width;
+			double Y2 = Y + Height;
 			GraphicsManager.SetColor(Color.White);
 			GL.BindTexture(TextureTarget.Texture2D, texture.ID);
 			GL.Begin(BeginMode.Quads);
